Dissertation Horror Game Version 1
Here is the first version of my dissertation project, a simple horror game where the player must find and collect 8 skulls around the various landmarks of the forest. There are no enemies, however there are several jumpscares in various locations. This may be worked on after I have finished University.